+++ internal/feed/cache/cache.go
@@ -24,6 +24,7 @@ const (
 type Impl interface {
 	cachedFeed(*feed.Feed) CachedFeed
 	transformTo(Version) (Impl, error)
+	cleanup(knownDescriptors map[feed.Descriptor]bool)
 	load(io.Reader) error
 	store(io.Writer) error
 	Version() Version