+++ pkg/util/util.go
@@ -21,3 +21,8 @@ func TimeFormat(t time.Time) string {
 
 	return t.Format(time.ANSIC)
 }
+
+// Days returns the number of days in a duration. Fraction of days are discarded.
+func Days(d time.Duration) int {
+	return int(d.Hours() / 24)
+}